requirements.txt
requests
bs4
discord.py
# discord.py==0.16.12
